Change university selector.

+++ main-ms/src/main/java/com/vidnichuk/isogj/api/dto/model/UserCompanyDto.java
@@ -1,18 +1,16 @@
 package com.vidnichuk.isogj.api.dto.model;
 
-import com.vidnichuk.isogj.api.model.Company;
-import com.vidnichuk.isogj.api.model.Position;
 
 public class UserCompanyDto {
-    private Long idUserCompany;
+    private int idUserCompany;
     private CompanyDto company;
     private PositionDto position;
 
-    public Long getIdUserCompany() {
+    public int getIdUserCompany() {
         return idUserCompany;
     }
 
-    public void setIdUserCompany(Long idUserCompany) {
+    public void setIdUserCompany(int idUserCompany) {
         this.idUserCompany = idUserCompany;
     }
